appEngine/python-skeleton-challenge/requirements.txt
google-endpoints==2.5.0
google-endpoints-api-management==1.5.1